+++ .vscode/settings.json
@@ -1,0 +1,8 @@
+{
+  "files.watcherExclude": {
+    "**/.git/objects/**": false,
+    "**/.git/subtree-cache/**": false,
+    "**/.hg/store/**": false,
+    "**/node_modules/*/**/node_modules/**": false
+  }
+}
